requirements.txt
discord.py=1.5.0
pip==20.2.3
setuptools==50.3.0